examples/elf : User can select the example which to test.
In wifi app, user can enter the input which test to run.
For recovery, press R or r.

Below is sample output for test.
This is WIFI App

Select Test Scenario.
        -Press R or r : Recovery Test
I'm RECOVERY main! create mkchildren task 9
[9] make_children_task
[10] normal_task
[11] assert_group_main_task
[12] 0th thread, normal thread
[13] 1th thread, normal thread
[14] 2th thread, Write another app space 0x80200100
